area_of_triangle

+++ 04.Input().py/10.area_of_triangle.py
@@ -1,5 +1,5 @@
-b =int(input("enter your triangle  breadth:"))
-h =int(input("enter your triangle height:"))
+b =float(input("enter your triangle  breadth:"))
+h =float(input("enter your triangle height:"))
 
 area =(1/2 * b*h)
 print("area of trinagle is:",area)
